pyproject.toml
[project]
name = "manage-breast-screening"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django (>=5.1.7,<6.0.0)",
  "dotenv (>=0.9.9,<0.10.0)",
  "gunicorn (>=23.0.0,<24.0.0)",
  "jinja2 (>=3.1.6,<4.0.0)",
  "whitenoise[brotli] (>=6.9.0,<7.0.0)",
  "nhsuk-frontend-jinja (>=0.2.0,<0.3.0)",
  "python-dateutil (>=2.9.0.post0,<3.0.0)",
  "psycopg[binary] (>=3.2.7,<4.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
factory-boy = "^3.3.3"
time-machine = "^2.16.0"
pytest-playwright = "^0.7.0"
ipdb = "^0.13.13"
axe-playwright-python = "^0.1.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "manage_breast_screening.config.settings_test"
python_files = "tests.py test_*.py *_tests.py"
addopts = "--doctest-modules"
markers = ["system: mark a test as a system test"]
